IFHAJF

+++ client.cpp
@@ -13,7 +13,7 @@ int main(){
     int s = socket(AF_BLUETOOTH, SOCK_STREAM, BTPROTO_RFCOMM); 
 
     addr.rc_family = AF_BLUETOOTH;
-    addr.rc_channel = 1; // may need to cast to uint_8
+    addr.rc_channel = (uint8_t) 1; // may need to cast to uint_8
     str2ba(BT_address, &addr.rc_bdaddr);
 
     int status = connect(s, (struct sockaddr* )&addr, sizeof(addr)); 
@@ -25,7 +25,7 @@ int main(){
     std::string command; 
 
     // begin CLI functionality
-
+    // write(s, command.c_str(), command.length());
     return 0; 
 }
